api/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
PyJWT>=2.0,<3.0
django-cors-headers>=3.7,<4.0